move price lookup functions out of app
